chore: move component tests files into a centralized components test dir
